+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Fixed
 
 * Fix zeroing pages when incremental savestates feature is disabled
+* Specifying a movie in commandline does not truncate the file anymore
 
 ## [1.3.2] - 2018-11-16
 ### Added